% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\title{survey}
\usage{
survey(filenames)
}
\arguments{
\item{filenames}{"C:/Data_Science/Text_Analysis/IPO_R&D_Raw.xlsx"}
}
\description{
This function return incidents
}
\examples{
survey("C:/Data_Science/Text_Analysis/IPO_R&D_Raw.xlsx")
}
