% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_score.R
\name{survey_score}
\alias{survey_score}
\title{Survey Score}
\usage{
survey_score(input)
}
\arguments{
\item{survey.Ticket.Country}{India}

\item{survey.Ticket.Assigned.Group}{SYMPHONY_WW_L2}

\item{survey.Product.Name}{Symphony}

\item{survey.Category.Tier1}{Troubleshoot}

\item{survey.Category.Tier3}{License}
}
\value{
data frame
}
\description{
This function return the satisfaction rating prediction
}
\examples{
 input <- data.frame(
 survey.Ticket.Country = "India",
 survey.Ticket.Assigned.Group="SYMPHONY_WW_L2",
 survey.Product.Name="Symphony",
 survey.Category.Tier1="Troubleshoot",
 survey.Category.Tier3="License"
 )
 OR
 input = "C:/Data_Science/SurveyPrediction/input.csv"
 SurveyPrediction::survey_score(input)
}
